big warning message in build output from million js
